requirements.txt
hikari==2.0.0.dev105
hikari-lightbulb==2.1.3
uvloop==0.16.0
psutil==5.8.0
colorama==0.4.4
uptime==3.0.1
tldextract==3.1.2
aioredis==2.0.1
pytz==2021.3
aiohttp==3.8.1